Some fixes to Tracing Path
Thank you Parker for helping me implement the game. I have patched a few issues with the game, there was a bug where if you exit to menu then return, the game become unplayable but it is fixed now (hopefully). I also fixed the issue with the progress bar's borders not showing, and moved the instructions to the "?" button.